+++ macro/hk.py
@@ -28,7 +28,7 @@ def initialize(frame, btn, runs=1):
   shortcut.add_hotkey(util.HOTKEY_TERMINATE, util.terminate)
   btn_start.config(state="disabled")
   frame_root.update()
-  run_dungeon(int(runs))
+  run_dungeon(runs)
   btn_start.config(state="active")
   frame_root.update()
 
@@ -135,8 +135,10 @@ def find_mobs(unit=util.UNIT_BLANK):
 def run_dungeon(runs=1):
   run_counter = 0
   while run_counter < runs:
+    util.check_reconnect(run_counter)
     run_counter += 1
     util.set_reset_status(False)
+    util.check_reconnect(run_counter)
     util.log_action(util.MSG_START_DG)
     util.log_run(run_counter)