+++ CHANGELOG.md
@@ -20,10 +20,13 @@ All notable changes to this project will be documented in this file.
 ## v5.31 Features
 
 - [Application] Added get mail feature
+- [Application] Added reconnect feature in Developmental Phase
 - [Dungeons] Added necessary log actions when moving to positions
+- [Application] Added reconnect check in Developmental Phase
 
 ## v5.31 Fixes
 
+- [Application] Disabled party and leader feature temporarily
 - [HK] Increased count threshold from 20 to 50 when waiting for the third boss
 - [HK] Changed slightly dashing position when coming back for final after doing battle mode
 - [HK] Fixed second boss sequence